# apps/api/openapi.json
{
  "components": {
    "schemas": {
...
      "CompileResponse": {
        "additionalProperties": false,
        "properties": {
          "compilerOutput": {
            "title": "Compileroutput",
            "type": "string"
          },
          "compilerOutputTruncated": {
            "title": "Compileroutputtruncated",
            "type": "boolean"
          },
          "diagnostics": {
            "items": {
              "$ref": "#/components/schemas/CompileDiagnostic"
            },
            "title": "Diagnostics",
            "type": "array"
          },
          "durationMs": {
            "title": "Durationms",
            "type": "integer"
          },
          "firmware": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Firmware"
              },
              {
                "type": "null"
              }
            ]
          },
          "sizes": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CompileSizes"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "enum": [
              "success",
              "error"
            ],
            "title": "Status",
            "type": "string"
          },
          "toolchain": {
            "$ref": "#/components/schemas/Toolchain"
          }
        },
        "required": [
          "status",
          "diagnostics",
          "sizes",
          "firmware",
          "compilerOutput",
          "compilerOutputTruncated",
          "toolchain",
          "durationMs"
        ],
        "title": "CompileResponse",
        "type": "object"
      },
...
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "ProjectCreate": {
        "additionalProperties": false,
        "properties": {
          "board": {
            "$ref": "#/components/schemas/BoardType",
            "default": "arduino-uno-r3"
          },
          "circuit": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Defaults to an empty circuit with the board only.",
            "title": "Circuit"
          },
          "code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Defaults to an empty sketch skeleton.",
            "title": "Code"
          },
          "description": {
            "$ref": "#/components/schemas/ProjectDescription",
            "default": ""
          },
          "name": {
            "$ref": "#/components/schemas/ProjectName"
          }
        },
        "required": [
          "name"
        ],
        "title": "ProjectCreate",
        "type": "object"
      },
...
      "ProjectDetail": {
        "additionalProperties": false,
        "properties": {
          "board": {
            "$ref": "#/components/schemas/BoardType"
          },
          "circuit": {
            "additionalProperties": true,
            "description": "Circuit document (CircuitDocument).",
            "title": "Circuit",
            "type": "object"
          },
          "code": {
            "title": "Code",
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "revision": {
            "description": "Current project revision (optimistic concurrency token).",
            "title": "Revision",
            "type": "integer"
          },
          "schemaVersion": {
            "title": "Schemaversion",
            "type": "integer"
          },
          "updatedAt": {
            "format": "date-time",
            "title": "Updatedat",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "board",
          "schemaVersion",
          "revision",
          "createdAt",
          "updatedAt",
          "code",
          "circuit"
        ],
        "title": "ProjectDetail",
        "type": "object"
      },
      "ProjectList": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ProjectSummary"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "ProjectList",
        "type": "object"
      },
...
      "ProjectUpdate": {
        "additionalProperties": false,
        "description": "Частичное изменение: отсутствующие (или null) поля не меняются.",
        "properties": {
          "circuit": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Circuit"
          },
          "code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Code"
          },
          "description": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProjectDescription"
              },
              {
                "type": "null"
              }
            ]
          },
          "name": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProjectName"
              },
              {
                "type": "null"
              }
            ]
          },
          "revision": {
            "description": "Revision the change is based on; a mismatch returns REVISION_CONFLICT.",
            "minimum": 1.0,
            "title": "Revision",
            "type": "integer"
          }
        },
        "required": [
          "revision"
        ],
        "title": "ProjectUpdate",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "MicroLab API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/projects": {
      "get": {
        "operationId": "listProjects",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectList"
                }
              }
            },
            "description": "Successful Response"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unexpected server error."
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication is not configured (AUTH_NOT_CONFIGURED)."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Database is unavailable or the development user is missing (DEV_USER_MISSING)."
          }
        },
        "summary": "List the current user's projects (most recently updated first)",
        "tags": [
          "projects"
        ]
      },
      "post": {
        "operationId": "createProject",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Source exceeds 256 KiB (SOURCE_TOO_LARGE)."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unexpected server error."
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication is not configured (AUTH_NOT_CONFIGURED)."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Database is unavailable or the development user is missing (DEV_USER_MISSING)."
          }
        },
        "summary": "Create a project",
        "tags": [
          "projects"
        ]
      }
    },
    "/api/v1/projects/{project_id}": {
      "delete": {
        "operationId": "deleteProject",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Project not found (PROJECT_NOT_FOUND)."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unexpected server error."
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication is not configured (AUTH_NOT_CONFIGURED)."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Database is unavailable or the development user is missing (DEV_USER_MISSING)."
          }
        },
        "summary": "Delete a project and its revisions",
        "tags": [
          "projects"
        ]
      },
      "get": {
        "operationId": "getProject",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Project not found (PROJECT_NOT_FOUND)."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unexpected server error."
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication is not configured (AUTH_NOT_CONFIGURED)."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Database is unavailable or the development user is missing (DEV_USER_MISSING)."
          }
        },
        "summary": "Get a project",
        "tags": [
          "projects"
        ]
      },
      "patch": {
        "operationId": "updateProject",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Project not found (PROJECT_NOT_FOUND)."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The project changed since the given revision (REVISION_CONFLICT)."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Source exceeds 256 KiB (SOURCE_TOO_LARGE)."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unexpected server error."
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication is not configured (AUTH_NOT_CONFIGURED)."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Database is unavailable or the development user is missing (DEV_USER_MISSING)."
          }
        },
        "summary": "Update a project (optimistic concurrency via revision)",
        "tags": [
          "projects"
        ]
      }
    },
    "/api/v1/projects/{project_id}/compile": {
      "post": {
        "operationId": "compileProject",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompileResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Project not found (PROJECT_NOT_FOUND)."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Source exceeds 256 KiB (SOURCE_TOO_LARGE)."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unexpected server error."
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication is not configured (AUTH_NOT_CONFIGURED)."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Database is unavailable or the development user is missing (DEV_USER_MISSING)."
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Time limit exceeded (COMPILATION_TIMEOUT)."
          }
        },
        "summary": "Compile the stored sketch of a project",
        "tags": [
          "projects"
        ]
      }
    },
    "/api/v1/projects/{project_id}/revisions": {
      "get": {
        "operationId": "listProjectRevisions",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectRevisionList"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Project not found (PROJECT_NOT_FOUND)."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unexpected server error."
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication is not configured (AUTH_NOT_CONFIGURED)."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Database is unavailable or the development user is missing (DEV_USER_MISSING)."
          }
        },
        "summary": "List stored revisions of a project (newest first)",
        "tags": [
          "projects"
        ]
      }
    },
    "/api/v1/projects/{project_id}/revisions/{revision}": {
      "get": {
        "operationId": "getProjectRevision",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "revision",
            "required": true,
            "schema": {
              "title": "Revision",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectRevisionDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "PROJECT_NOT_FOUND or REVISION_NOT_FOUND."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unexpected server error."
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication is not configured (AUTH_NOT_CONFIGURED)."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Database is unavailable or the development user is missing (DEV_USER_MISSING)."
          }
        },
        "summary": "Get a stored revision of a project",
        "tags": [
          "projects"
        ]
      }
    },
    "/api/v1/projects/{project_id}/validate": {
      "post": {
        "operationId": "validateProject",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CircuitValidationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Project not found (PROJECT_NOT_FOUND)."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unexpected server error."
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication is not configured (AUTH_NOT_CONFIGURED)."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Database is unavailable or the development user is missing (DEV_USER_MISSING)."
          }
        },
        "summary": "Validate the stored circuit of a project",
        "tags": [
          "projects"
        ]
      }
    }
  }
}